requirements.txt
googlesearch-python
pyyaml